Decompose the application.
